tests/test_models.py
# -*- coding: utf-8 -*-
"""Model unit tests."""
import arrow
import pytest

from pyxword_contest.user.models import Role, User

from .factories import UserFactory


@pytest.mark.usefixtures('db')
class TestUser:
    """User tests."""

    def test_get_by_id(self):
        """Get user by ID."""
        user = User('foo', 'foo@bar.com')
        user.save()

        retrieved = User.get_by_id(user.id)
        assert retrieved == user

    def test_created_at(self):
        """Test creation date."""
        start = arrow.utcnow()
        user = User(display_name='foo', email='foo@bar.com')
        user.save()
        assert bool(user.created_at)
        assert start < user.created_at < arrow.utcnow()

    def test_password_is_nullable(self):
        """Test null password."""
        user = User(display_name='foo', email='foo@bar.com')
        user.save()
        assert user.password is None

    def test_factory(self, db):
        """Test user factory."""
        user = UserFactory(password='myprecious')
        db.session.commit()
        assert bool(user.display_name)
        assert bool(user.email)
        assert bool(user.created_at)
        assert user.is_admin is False
        assert user.active is True
        assert user.check_password('myprecious')

    def test_check_password(self):
        """Check password."""
        user = User.create(email='foo@bar.com', password='foobarbaz123')
        assert user.check_password('foobarbaz123') is True
        assert user.check_password('barfoobaz') is False

    def test_roles(self):
        """Add a role to a user."""
        role = Role(name='admin')
        role.save()
        user = UserFactory()
        user.roles.append(role)
        user.save()
        assert role in user.roles
