requirements/prod.txt
# Everything needed in production

# Flask
Flask==0.12.2
MarkupSafe==1.0
Werkzeug==0.12.2
Jinja2==2.9.6
itsdangerous==0.24
click==6.7
marshmallow==2.13.5
flask-marshmallow==0.8.0
Flask-Admin==1.5.0

# Api
flask-apispec==0.4.0
flask-restful==0.3.6

# Database
Flask-SQLAlchemy==2.2
psycopg2==2.7.2
SQLAlchemy==1.1.11

# Migrations
Flask-Migrate==2.1.0

# Forms
Flask-WTF==0.14.2
WTForms==2.1

# Deployment
gunicorn>=19.1.1

# Webpack
flask-webpack==0.1.0

# Auth
Flask-Login==0.4.0
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==2.4.1

# Caching
Flask-Caching==1.3.2

# Data sanitization
Bleach==2.0.0

# JSON serialization (used only for the logs)
ujson==1.35

# E-mail shipment
flask-mail==0.9.1

# Debug toolbar
Flask-DebugToolbar==0.10.1

# Logging & error reporting
raven[flask]==6.1.0

# Request threshold
Flask-Limiter==0.9.5

# IPython shell
flask-shell-ipython

arrow==0.10.0
Markdown==2.6.8
python-slugify==1.2.4
PyYAML==3.12
SQLAlchemy-Utils==0.32.14

# Deployment
uwsgi

# Crosswords
git+git://github.com/rooterkyberian/acrossword.git#egg=acrossword
